Updated mediawiki/extensions Project: mediawiki/extensions/ArticleFeedbackv5  a372d393bdbf7eb15bb3cc3657e0cfb7df8f0d7a
(bug 38474) Mixed directionality of AFT

While the bug report saw it as a RTL issue, it could also occur on LTR.
If the content preceding the AFTv5 box is floated (either side) and small enough for AFT to fit next to it, it will.
Giving AFTv5 its own 100% witdh will make sure it doesn't sit next to floated elments.

Bug: 38474
Change-Id: I2f349df5c848bf36497ad383623b0daa71a2d3fb