Updated mediawiki/extensions Project: mediawiki/extensions/TemplateData  708bbad12d54c4d7186896670b93d28a37bb5580
Implement Sets

Change-Id: I93a0ba7287fec6e5566994d03dff1d8eaeeb8430

tests: Reformat tests, evaluate status in addition to I/O

Previously we only compared output, we now evaluate
status as well.

For example test {} => {} was misleading since the reason
it roundtripped was because it is *invalid* and we replace
invalid blobs with an empty object in blob storage.

But in the API this'll never get used because getStatus
will have a fatal error in it. Which overrides the data.
The tests were previously looking in the data whilst ignoring
the return status.

Change-Id: I9809f126c1615c9bc9c28ce2d8bb7953058cbf6e