Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  cf3d19925f312ee42e7f32137f5f80c468496ce1
(FR #908) Brazil requires extra personal infos

The fields will show up on every GC payment method, when the donor is geolocated
to Brazil, but values are only mandatory when donating using Boletos.  If we
decide to implement Brazil BT, we will need the fields to be mandatory there as
well.

Change-Id: Ib0b3de3d76e1a98a253cf9a9f4f88dc0ec141d14