Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  dc06b430bf68c6e900eb8724316608abd7c31775
Adding a maintenance script to resend email confirmation emails.

Instead of Special:ConfirmEmail, the links direct to Special:MergeAccount
which has been modified to confirm an email address when a token is present.
This allows us to start the migration process and get the users straight into
the account migration flow without having to modify already translated
messages and requiring them to manually enter the workflow.

Change-Id: Ia6ee70420cb87f523339cfd5a7b391a1fcedb6a2