Updated mediawiki/extensions Project: mediawiki/extensions/DataValues  7b7ffd7cc79847a952c55eba16058b65d2d415fd
(bug 48145) Adding calendar hint message

Message informing about the calendar that is detected automatically when specifying a date.
The message will be shown only when the specified date lies within a time frame when multiple
calendars had been in use. Since only Gregorian and Julian calendar are implemented at the
moment, this is specifically implemented according to those calendars.
In addition to the plain message, a links allows switching between the currently implemented
calendars.

Change-Id: I8b3d2bb6c04b01fc987c2c4ed9c0bd406a824134