Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  91594db5c526b4de45a23718fa28a7bfbdd2309e
Add Ajax-y interface for Special:NovaAddress

Continuing along the trend!

Refactored some things so it's easier to do some of the dependency and general
things that more than one interface is going to need to do (like notifications
and having a base class), and modified the Instance module to make those
things work.

Change-Id: Ic5f264707b3583bf61031d0d6f2774792b5ba8d0