Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  8ffb84a9b0170497ec88a6485c15d93cb8f31ca6
Added test config and bootstrap for the DataModel component

This commit fixes most of the dependency issues where DataModel code
was still using things from core or wikibase lib. Some issues
are however more serious and fixing them is definitly out of scope
for this commit, so the tests invoking this code have been excluded from
the WikibaseDataModelWithoutMW testsuite for now.

Change-Id: I44397a6b6384cb0210e7a08a552bea90515e4716