Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  26e976567a3b6540cda1d3364d0c61805133d312
Fix a bug with trimming 'local-' from the requested service group name.

If a user is overly clever and included $wgOpenStackManagerServiceGroupPrefix
at the start of the service group they created, we trimmed it in most
cases but left it in for the $homedir setting.  That caused
various confusions.  Now we trim it in all cases.

This should fix bug 48699.

Change-Id: I0f774afb40ac95eb52de7960b02033bc61178f3a