Updated mediawiki/extensions Project: mediawiki/extensions/ExternalArticles  c260ee2d7bc3a32cd74cd7784644c288def95afa
Add a message to the edit form when external text is loaded

This change adds a message, with 'success' class (so green by
default), above the page edit form when an external article's
text has been loaded. This message shows the external URL that
the text has been loaded from.

Change-Id: Ib2700ba24e055816ecd0809cb4d0a71d0d661326