Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  d542db6e36c790c645e0fbcadb89996ec442537b
Story 770: Add notifications (user) icon

When Echo is loaded, the icon shows the number of unread notifications
and is a link to Special:Notifications. When user is logged out, shows a
CTA drawer.

Also, removes the search button on special pages in alpha because there
is no room for it anymore.

Change-Id: I7fb40be2ff2fd6988dc2f64110a76a5867300ba2