Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  1e80984c6755bfcae27098eea47de8d74314e939
Allowing the OrderID for Adyen to be Resurrected

Otherwise nothing sets it and we lose it!

Change-Id: Iee81d1a78b7d0883139f1b6abe3e92a2d5703b2a

Changes for SmashPig/Adyen!

When a payment returns from Adyen it should be in the pending
state because we always will need to do further things to it.

And... we need to add metadata to stomp messages such that
smashpig doesn't freak out when it attempts to deserialize.

Change-Id: Ia733b4fb1cabab2ab34e084e96c212d1675538ae