Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  058da08221ceb50f7b6ea6ffad37985387005f8c
Force page DOM generation for selective serialization

The ParsoidCacheRequest normally sends a Cache-Control: only-if-cached header
to Varnish, which prevents a page from being generated if not in cache. For
selective serialization we need the DOM in any case, so this patch adds an
option to disable the only-if-cached behavior in ParsoidCacheRequest and uses
that when retrieving a DOM for selective serialization.

Change-Id: If5568fb465fabde52cb9a00763f323b00415a066