Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  187c996268414213564231732c9713e1aceccf6d
Remove DataModel component as it now resides in the WikibaseDataModel git repo

The new repo is located at
https://gerrit.wikimedia.org/r/p/mediawiki/extensions/WikibaseDataModel.git

This means that Wikibase now needs the WikibaseDataModel code to be present. If
it is checked out as MediaWiki extension, it will be automatically loaded. Composer
support has also been added.

The removed test was misplaced and depended on WikibaseDatabase component code,
while WikibaseDatabase is now no longer automatically loaded. It previously was
since it was required by WikibaseQueryEngine, but this component has already been
moved out a few days ago.

Change-Id: I36964757aac2276115863c705079607257e175b2