Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  779cd70f59b3c497a19768702b8e956e4b995d4d
Use SnakConstructionService to wrap SnakFactory.

This allows the SnakFactory to operate without the need to
know about Properties and their DataValues.

NOTE: this requires Ifab91f9c633 to be merged into WikibaseDataModel.

NOTE: the name SnakCreationService isn't sexy (but SnakFactory was taken,
and SnakBuilder would be misleading). I'm geatful for suggestions...

Change-Id: Ib042d74383ea25d257a6736bc181a9bd1951989d