Updated mediawiki/extensions Project: mediawiki/extensions/Score  0fc78bc4762d2bec5b89b91255076c5f486d4bc4
Make score work with multi-page files

Score was looking for files of the form "foo-1.png", "foo-2.png"
where really it should have been looking for "foo-page1.png".
(Someone should double check that it isn't just my install of
lilypond being broken or something).

Also there was a variable name mismatch in the multipage handling
code.

I'm not sure if it really is appropriate to have "Page 1" in the
alt text. I can't imagine what a blind person would do with that
info, however I left it there. I also added the page number to
title, as its mildly useful as a tooltip.

As an aside, I'm not sure if it would make sense to add a sanity
check somewhere - only copy over a max 5000 pages or something.

Change-Id: I09f1fca86a38655dfc1583835c3e59823a08653c