Updated mediawiki/extensions Project: mediawiki/extensions/SemanticMediaWiki  3110db4b3edede032ae5906baa5b43a6fbfd9253
Display an introductory text for predefined properties

This is for users who barely know SMW (and for all others as well). It
will display an introductory text for all predefined properties and
links to smw.org allowing for easier access to available online resources.

It is further possible to maintain an individual text component for each
predefined property using the following message key
'smw-pa-property-predefined' + <internal property key>.

## Example "Query size"

'smw-pa-property-predefined' + '_asksi' = smw-pa-property-predefined_asksi'

"Query size is a predefined property (also known as special property)
which collects the number of conditions used in a query. It is a
built-in property that comes with additional administrative privileges
but can be used just like any other user-defined property."

## Red links and predefined properties

Using the TitleIsAlwaysKnown hook to override the "known" status which
enables predefined properties to be shown as blue links.

TitleIsAlwaysKnown hook was introduced in MW 1.20 which means that for 1.19
predefined properties are still displayed as red links (that's the price
to be paid for people running an old MW). Of course you can
combine LinkBegin + BeforeDisplayNoArticleText to shoulder also MW 1.19
but that's too much effort to justify to make it workable in MW 1.19
as well.

One could extend SMWWikiPageValue to set the link options parameter
$options[] = 'known' which would solve the issue for MW 1.19
but that's not part of this change.

Bug: 23691

Change-Id: I029716a67302962e1b6e2d9a769c150b63d7da74