Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  21e1e7d82ab6cf012ebb573738b8ea375a1f9634
Add a lang member to the wiki config and compare CONTENTLANGUAGE with that

The interwiki prefix is the same as the content language normally, but in
production we are using db names instead. This patch adds a lang member to the
wiki config and changes the API debug code to compare CONTENTLANGUAGE with it.

Change-Id: I459da83c56bb7af21db55f5ecb3b85dee2dbbcfe