Updated mediawiki/extensions Project: mediawiki/extensions/DataValues  ac6711a08822fcc92d89cf28c2b5e8a969719c1a
Implements autoActive() for listrotator widget

Allows to check whether the "auto" button is currently active on the widget.
Also removed the "auto" option in this change since the option has no effect on the widget even
though it was documented and has a default value. "auto" is just always active on initialization.

Change-Id: I57a6ad92c010a1e577b2418a6fabf63652ad6def