Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  454f87a2edb699d76df48910eef671c39a16b72c
Bug 50003: Fix template and image updates

I verified that the update requests properly reach the Parsoid backends, but
those currently just re-render the page without updating transclusion or file
expansions. The reason was that the ParserService did not properly clear the
caches.

I also took the opportunity to rename images / imageCache to files / fileCache
for consistency.

Change-Id: I0084c8b28551a878e71a390da4e9b8fa9c9c1a53