Updated mediawiki/extensions Project: mediawiki/extensions/VisualEditor  e30c346d1cc32efb36bb1085cf7ce0ccfb94170f
Fix for edit links, RTL wikis with LTR interface

Added both bidi-isolation and directionality. Since the directionality
has to come from the user interface (and not the wiki dir itself) I had
to read the $('body') direction. The other option could be adding a
mw.config.get().userLanguageDir that picks up the interface directionality
on top of the wiki directionality.

Bug: 50543
Change-Id: I9a00e9545a46f13750ab0d118b23d05573ee1a2e