Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  e20fec628e0532d4ea974a006af3adec47f1f6fc
Bug 50636: Don't extrapolate DSR on change

The WikitextSerializer's emitSeparator method was using text node lengths to
extrapolate a previousSibling's DSR, even if that text node was changed. This
led to an invalid extrapolated dsr if that text was modified.

This patch adds a test for the (new) children-modified diff marker on the
parentNode and does not extrapolate the DSR in that case.

Fixes four selser tests in the process.

Change-Id: Iad67bac7a9ba890cbd3816d3ff3295b23b34be30