Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  d8c1f40f66338e3d12be779be18284a9cddae192
Cog size adjustments

SVG and PNG versions have been modified to be of the same dimensions.
The sprite now is a 14x14px cog with an additional transparent pixel
above and below. resulting in two 14x16px areas. The added margin solves
the problem of image being cropped when zooming.

Bug: 50742
Change-Id: Iab4f578bdb1ea55c51eb528527fc38ce44c7be69