Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  d08effc21e4deebbdc412d08dd4859db89fafa62
Label and description serialization now accepts LanguageFallbackChain as the language option

Labels and descriptions in all serialized formats will get a new field 'source-language',
to indicate the language where this label was fetched.

If the target format doesn't have keys, one new extra field 'for-language' is added, to
hold the language which is used as keys "normally".

Change-Id: Iec62d6817e3007e1982db65b5031ec09ecda3374