Updated mediawiki/extensions Project: mediawiki/extensions/CloseWikis  b2749e1f57fe5979062a87e7fc232a3d67f097f7
Renaming files to standardise files that are part of special pages.

Old names such as "CloseWikis.page.php" is not very descriptive at
all, and this commit moves the files that generate the Special pages
to a better naming standard such as SpecialClosewiki.php. The main
extension file is updated to reflect this change.

Change-Id: I871b3e4a466f270c92c0c1607118cd8580a4a040