Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  c1afe8b638a1cf0ee5d56253108a800a5582293d
Check if Diff is an array before accessing field

This is to increase the robustness with regards to
possible values given by the Change object. Although this
should never happened, it seemed to have caused the
Disptacher choking on July 12/13, 2013, so we just check
here one more time whether the variable is an Array
before accessing its field. A full analysis of the actual
cause of the event is still to happen, thought.

Change-Id: Ibf1b1f01d82cc2e2febabbe2cebbace3508ee5d6