Updated mediawiki/extensions Project: mediawiki/extensions/TranslationNotifications  c4b8c9a4fb4f2e704acc90397645158273ae437b
If the local user has the bot userright, mark the edit as bot.

In I0fbe0646, I submitted a patch that fixed this for the publishInOtherWiki
function. This patchset addresses the publishHere function,
which is used for notifications on the same wiki as the translation
(e.g. Meta translation notified on Meta user talk).

The function now checks if the user has the "bot" right, and if so modify the
bitflags to include EDIT_FORCE_BOT. We could theoretically add the flag even
if the user did not have the "bot" right, but this allows the local wiki to
choose whether they want it marked as a bot or not.

The user receiving the message will always get the new notifications message
since EDIT_FORCE_BOT and EDIT_MINOR will never both be set.

Bug: 41637
Change-Id: I68202241fd83b652fe5fc2cf33f5080ec5961d2f