Updated mediawiki/extensions Project: mediawiki/extensions/LiquidThreads  811adb76e48f2f52060bdd63eea2b693e876ba77
Fix breakage from Ie33446a2

LQT was relying on the presence of a 'watchlist' key on the personal URLs array.
With Ie33446a2 this would no longer occur if the user had no 'viewmywatchlist' permission,
therefore the 'New messages' link would not appear at the top and a PHP Notice would be generated.

Change-Id: Icb81016da76911599c092a93cea3558d49d5293d