Updated mediawiki/extensions Project: mediawiki/extensions/Echo  b9e5b2f64a9758c4dfa9be47f46e3b4a8d318c9d
(bug 51640) wgUserNewMsgRevisionId is null with only one rev

wgUserNewMsgRevisionId is null with only one revision and that revision is
not viewed, this makes it not reliable for determining if there is new
message.  We just switch to check against $user->getNewtalk() instead.

We can still use wgUserNewMsgRevisionId to generate diff links if desired

Change-Id: I4cf50a944aada03151bd17f3610bd59b3bfb2bf2