Updated mediawiki/extensions Project: mediawiki/extensions/Maps  26a0bf7b63fd88a01f03a2552c38374349aedc23
Added method to toggle kml (geoxml3) layers on and off.

Users will now be presented with a set of checkboxes for each
kml file loaded on the map. clicking the checkboxes will hide or show the
corresponding kml data.

Also added feedback to user in the form of an error message when
failure occurs in parsing kml files.

Change-Id: Iade1c305e70d9aa98c3927f52794a5009056ad88