Updated mediawiki/extensions Project: mediawiki/extensions/CodeReview  055511c44211e5f08d1440488820725d3e9d1a46
Add Renameuser support to CodeReview.

Unfortunately this requires a schema change since the code_authors table
originally stored only the user's name instead of both the name *and* ID,
as the convention has been pretty much since the beginning of time.

New maintenance script, populateCaUserColumn.php, can be used to populate
the column when there are pre-existing entries (as is the case with each
and every wiki already using CodeReview, such as MediaWiki.org).

Also bumped the version number.

Change-Id: I4e3dc200de61c55a4d031b36048700e1817aa088