Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  910167b58f73054d1d2ce7d39573798fcd10b046
Register NS_CAMPAIGN at file level in extension's entry point

The point at which $wgExtensionFunctions are executed is too late for
registering additional namespaces. Defering the registration of the namespace
until then can lead to unpredictable results, as documented in bug 45031.

The reason for deferring namespace registration to a CanonicalNamespaces hook
handler is to wait for user configurations to load before deciding whether or
not to register the namespace. Because UploadWizard registers the namespace
unconditionally, it is not necessary to defer the registration of the namespace
at all; it can be done at file scope in the extension's entry point.

Bug: 45031
Change-Id: I2fb2bcaa2108dd40a5374505ee5a9d8845e6e1d8