Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  856e6cdbdacf6779068c0309a7dc0d4797cd771a
(Bug 50982) Fix for empty selser changes arrays.

Matching against the parsoid options array, when defined, was causing
the selser test to be filtered out.

The fix checks if either "selser" or "wt2wt" is present in the array,
future proofing against a possible selser option.

Also regenerates the selser changes files and blacklist.

Bug: 50982
Change-Id: Ic27328d03288d0cc70c18682ab696654227f086a