Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  80ee2d020896b2d946a08e065e7251e510e04fb4
Parse the input list using a parser function

Users would use {{#target:User talk:Example|domain}} to add a subscription.
The domain part is optional and will result in a local delivery if not provided.

$wgNamepsacesToExtractLinksFor was renamed to $wgNamespacesToPostIn, and now controls
which namespaces we can post in. This checking is now done when the job executes so each
wiki can control their own settings. Basic duplicate checking is still done when parsing
the spamlist.

A dependency is added upon the SiteMatrix extension for any global deliveries.

Change-Id: I6cf785a1e89cad6f30b6abf7b00d5a9845e66e4b