Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  17e3479701f4eeacc2c0b6f71b72efc1f757cf86
Skip redirects in forceSearchIndex run by id.

Note that we don't skip redirects when run by timestamp because that is
assumed to be for an outage.
Note that we now properly skip doing anything when a page redirects to a
page that his been deleted.
Finally also note that if you happen to get a whole batch of redirects
all at once this doesn't cause the process to stop prematurely.  Yay.

Change-Id: I54ccb8b062ce97350e8a9f688510ea575526bd9d