Updated mediawiki/extensions Project: mediawiki/extensions/Parsoid  81ab0f9f1d9a5a0830f1d14a98e10a959d173620
Fixes PreHandler nested <pre> problem.

When running the tests with just --html2html, a bunch of unexpected
passed show up. Upon investigation, it seems the tokenTransformers per
key array contained duplicates (ex. 2.95 occurred twice), causing extra
newlines in the output, and such.

The source of the duplicates is the PreHandler. It seems confused by
wikitext like `<pre>\n string\n</pre>` where a space before string would
normally indicate indentation and kick off pre wrapping. This is
undesirable because we're already inside a pre. It ends up emitting nested
<pre>s. Once that happens, the ParagraphWrapper gets all out of sorts
and tries to add transformers out of order. In the test suite, this starts
at "HTML-pre: 2: indented text".

The change asserts uniqueness in the array to avoid this situation again.

The blacklist was rewritten with a bunch test fixes. Microdata
selser tests generate differently as a result of the changes but continue
to fail.

Bug: 52046
Change-Id: I8dd9b197be92de1cab4408a14f1dad302f4b9944