Updated mediawiki/extensions Project: mediawiki/extensions/CoreEvents  94fd3cc35dd6d52f45b50d9711ea2ea4d0eb739d
Update for new event schema

This change fulfills a request for additional datapoints for HTTP / HTTPS
logging. It adds the user's authentication status, mobile mode (if using the
mobile site) and GeoIP-derived city data. It also increases the rate at which
we log for logged-in users. The schema changes are captured by this diff:
http://tinyurl.com/5712722 (link points to MetaWiki).

Change-Id: Ide8469db2124afbf1cc6c40fbe161f23ebd1addb